python-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
torch>=2.2.0
git+https://github.com/openai/CLIP.git
Pillow>=10.0.0
requests>=2.31.0
python-multipart>=0.0.9
